requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
arrow==1.2.3
astunparse==1.6.3
async-timeout==4.0.2
attrs==23.1.0
braceexpand==0.1.7
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
discord-ext-bot==1.0.1
discord.py==2.3.0
frozenlist==1.3.3
ics==0.7.2
idna==3.4
import-expression==1.1.4
jishaku==2.5.1
line-profiler==4.0.3
multidict==6.0.4
python-dateutil==2.8.2
requests==2.31.0
six==1.16.0
TatSu==5.8.3
urllib3==2.0.3
yarl==1.9.2
